refactor: useCardNumber 관심사 별로 분리
