+++ hooks/src/lib/hooks/useCardNumber/constants.ts
@@ -1,0 +1,12 @@
+export const CARD_BRANDS = {
+  Visa: "Visa",
+  MasterCard: "MasterCard",
+  Diners: "Diners",
+  AMEX: "AMEX",
+  UnionPay: "UnionPay",
+  Unknown: "Unknown",
+} as const;
+
+export const ERROR_MESSAGES = {
+  number: "카드 번호는 숫자만 입력 가능합니다.",
+};
