+++ hooks/src/lib/hooks/useCardNumber/types.ts
@@ -1,0 +1,3 @@
+import { CARD_BRANDS } from "./constants";
+
+export type CardType = keyof typeof CARD_BRANDS;
